client,cli: decouple user methods from common
